+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Application/Authentication/ChangePassword/Command.cs
@@ -1,0 +1,3 @@
+namespace TimeSheet.Modules.EmploymentManagement.Application.Authentication.ChangePassword;
+
+public sealed record Command(string CurrentPassword, string NewPassword);