+++ EMS/src/TimeSheet.Modules.EmploymentManagement.Application/Authentication/ChangePassword/Validator.cs
@@ -1,0 +1,19 @@
+using FluentValidation;
+
+namespace TimeSheet.Modules.EmploymentManagement.Application.Authentication.ChangePassword;
+
+public sealed class Validator : AbstractValidator<Command>
+{
+    public Validator()
+    {
+        RuleFor(x => x.CurrentPassword)
+            .NotEmpty();
+
+        RuleFor(x => x.NewPassword)
+            .NotEmpty();
+
+        RuleFor(x => x)
+            .Must(x => !string.Equals(x.CurrentPassword, x.NewPassword, StringComparison.Ordinal))
+            .WithMessage("newPassword must be different from currentPassword.");
+    }
+}